transition to v1alpha2 skeleton
Without retaining any unused v1alpha1 content, transition the project to
a v1alpha2 skeleton. The client, controller, and sidecar are largely
featureless, based on a fresh kubebuilder project.

The Kubernetes client API is also a featureless scaffold based on a
fresh kubebuilder project. All COSI resource Kinds are accounted for,
but no COSI-specific fields are defined in spec or status definitions.

Similarly, gRPC protobuf files have core definitions and placeholders
for RPC calls without field definitions needed for COSI use.

The intent of the skeleton is that API, gRPC, controller, and sidecar
functionality can be added as needed as features are added. This will
facilitate deeper discussions in the future on small implementation
decisions.

Signed-off-by: Blaine Gardner <blaine.gardner@ibm.com>